requirements.txt
alpaca-py>=0.13.0
yfinance>=0.2.36
pandas>=2.0.0
numpy>=1.24.0
vectorbt>=0.26.0
schedule>=1.2.0
python-dotenv>=1.0.0
requests>=2.31.0
